Add deleting machines sensor
